Rename kas-core::event::cx::cx_shell -> platform
